+++ storedimo/database/migrations/2025_04_30_123939_add_imagen_producto_to_productos.php
@@ -13,9 +13,15 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('productos', function (Blueprint $table) {
-            $table->longText('imagen_producto')->nullable()->after('id_tipo_persona');
-        });
+        // Schema::table('productos', function (Blueprint $table) {
+        //     $table->longText('imagen_producto')->nullable()->after('id_tipo_persona');
+        // });
+
+        if (Schema::hasTable('productos') && !Schema::hasColumn('productos', 'imagen_producto')) {
+            Schema::table('productos', function (Blueprint $table) {
+                $table->longText('imagen_producto')->nullable()->after('id_tipo_persona');
+            });
+        }
     }
 
     /**
@@ -25,8 +31,14 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('productos', function (Blueprint $table) {
-            $table->dropColumn('imagen_producto');
-        });
+        // Schema::table('productos', function (Blueprint $table) {
+        //     $table->dropColumn('imagen_producto');
+        // });
+
+        if (Schema::hasTable('productos') && Schema::hasColumn('productos', 'imagen_producto')) {
+            Schema::table('productos', function (Blueprint $table) {
+                $table->dropColumn('imagen_producto');
+            });
+        }
     }
 };
